refactor: implement client abstraction and update indexer architecture
- Add client abstraction layer for ENS and UNI indexers
- Update governor implementations to use new client pattern
- Refactor indexer initialization to support multiple protocols
- Improve separation of concerns between protocol-specific logic and shared infrastructure
