+++ apps/indexer/src/indexer/ens/governor.ts
@@ -1,80 +1,58 @@
-import { Account, Address, Chain, Client, Transport } from "viem";
-import { getBlockNumber, readContract } from "viem/actions";
+import { ponder } from "ponder:registry";
 
+import {
+  proposalCanceled,
+  proposalCreated,
+  proposalExecuted,
+  voteCast,
+  delegateChanged,
+  delegatedVotesChanged,
+} from "@/eventHandlers";
+import { DaoIdEnum } from "@/lib/enums";
 import { Governor } from "@/interfaces/governor";
-import { ENSGovernorAbi } from "./abi";
+import { dao } from "ponder:schema";
 
-export class ENSGovernor<
-  TTransport extends Transport = Transport,
-  TChain extends Chain = Chain,
-  TAccount extends Account | undefined = Account | undefined,
-> implements Governor
-{
-  private client: Client<TTransport, TChain, TAccount>;
-  private abi: typeof ENSGovernorAbi;
-  private address: Address;
+export function GovernorIndexer(governor: Governor) {
+  const daoId = DaoIdEnum.ENS;
 
-  constructor(client: Client<TTransport, TChain, TAccount>, address: Address) {
-    this.client = client;
-    this.address = address;
-    this.abi = ENSGovernorAbi;
-  }
+  ponder.on(`ENSGovernor:setup`, async ({ context }) => {
+    const votingPeriod = await governor.getVotingPeriod();
+    const quorum = await governor.getQuorum();
+    const votingDelay = await governor.getVotingDelay();
+    const timelockDelay = await governor.getTimelockDelay();
+    const proposalThreshold = await governor.getProposalThreshold();
 
-  async getQuorum(): Promise<bigint> {
-    const blockNumber = await getBlockNumber(this.client);
-    const targetBlock = blockNumber - 10n;
-    return readContract(this.client, {
-      abi: this.abi,
-      address: this.address,
-      functionName: "quorum",
-      args: [targetBlock < 0n ? 0n : targetBlock],
+    await context.db.insert(dao).values({
+      id: daoId,
+      votingPeriod,
+      quorum,
+      votingDelay,
+      timelockDelay,
+      proposalThreshold,
     });
-  }
+  });
 
-  async getProposalThreshold(): Promise<bigint> {
-    return readContract(this.client, {
-      abi: this.abi,
-      address: this.address,
-      functionName: "proposalThreshold",
-    });
-  }
+  ponder.on(`ENSGovernor:VoteCast`, async ({ event, context }) => {
+    await voteCast(event, context, daoId);
+  });
 
-  async getVotingDelay(): Promise<bigint> {
-    return readContract(this.client, {
-      abi: this.abi,
-      address: this.address,
-      functionName: "votingDelay",
-    });
-  }
+  ponder.on(`ENSGovernor:ProposalCreated`, async ({ event, context }) => {
+    await proposalCreated(event, context, daoId);
+  });
 
-  async getVotingPeriod(): Promise<bigint> {
-    return readContract(this.client, {
-      abi: this.abi,
-      address: this.address,
-      functionName: "votingPeriod",
-    });
-  }
+  ponder.on(`ENSGovernor:ProposalCanceled`, async ({ event, context }) => {
+    await proposalCanceled(event, context, daoId);
+  });
 
-  async getTimelockDelay(): Promise<bigint> {
-    const timelockAddress = await readContract(this.client, {
-      abi: this.abi,
-      address: this.address,
-      functionName: "timelock",
-    });
-    return readContract(this.client, {
-      abi: [
-        {
-          constant: true,
-          inputs: [],
-          outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-          payable: false,
-          stateMutability: "view",
-          type: "function",
-          name: "getMinDelay",
-        },
-      ],
-      address: timelockAddress,
-      functionName: "getMinDelay",
-    });
-  }
+  ponder.on(`ENSGovernor:ProposalExecuted`, async ({ event, context }) => {
+    await proposalExecuted(event, context, daoId);
+  });
+
+  ponder.on(`ENSToken:DelegateChanged`, async ({ event, context }) => {
+    await delegateChanged(event, context, daoId);
+  });
+
+  ponder.on(`ENSToken:DelegateVotesChanged`, async ({ event, context }) => {
+    await delegatedVotesChanged(event, context, daoId);
+  });
 }
