+++ apps/indexer/src/indexer/ens/index.ts
@@ -1,2 +1,3 @@
 export * from "./abi";
+export * from "./client";
 export * from "./governor";
