+++ apps/indexer/src/indexer/index.ts
@@ -1,2 +1,1 @@
 export { ERC20Indexer } from "./erc20";
-export { GovernorIndexer } from "./governor";
